+++ src/Main.java
@@ -1,9 +1,11 @@
+import database.DataProvider;
+
 import java.io.FileNotFoundException;
 
 public class Main {
     public static void main(String[] args) throws FileNotFoundException {
         DataProvider dataProvider = new DataProvider();
         dataProvider.initialize();
-        new Searcher(dataProvider.getDataContainer()).run();
+        new UserInterface(dataProvider.getDataContainer()).run();
     }
 }
